backend/pyproject.toml
[project]
name = "le_agent"
version = "0.0.1"
description = "General agent framework with LangGraph"
authors = [
    { name = "merlinzhu", email = "merlinzhu@lexin.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = "==3.12.2"
dependencies = [
    "colorlog>=6.9.0",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "json-repair>=0.41.1",
    "langchain-community>=0.3.21",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.5",
    "langchain-litellm>=0.2.1",
    "langchain-openai>=0.3.14",
    "langgraph>=0.3.31",
    "litellm>=1.72.4",
    "markdownify>=1.1.0",
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
    "readabilipy>=0.3.0",
    "sse-starlette==2.1.0",
    "uvicorn>=0.34.2",
    "langgraph-cli[inmem]>=0.3.3",
    "duckduckgo-search>=8.1.1",
    "playwright>=1.53.0",
    "ddgs>=9.0.0",
    "crawl4ai>=0.7.1",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools]
packages = ["nova"] # 新增 "nova"
[tool.setuptools.package-dir]
"nova" = "nova"
[tool.setuptools.package-data]
"*" = ["py.typed"]
